+++ src/components/WithFacelift.tsx
@@ -11,9 +11,24 @@ import {
   StylesProvider as MuiStylesProvider,
 } from '@material-ui/core/styles'
 import { PreviewStyles } from '../styles/PreviewStyles'
-import { useFaceliftSettings } from '../index'
+import { useFaceliftState } from '../index'
 
 import type { ThemeOptions as MuiThemeOptions } from '@material-ui/core/styles'
+
+/** @TODO
+  api.addNotification({
+    id: 'update',
+    link: '/settings/about',
+    content: {
+      headline: `Storybook Y is available!`,
+      subHeadline: `Your current version is: X`,
+    },
+    icon: { name: 'book' },
+    onClear() {
+      store.setState({ dismissedVersionNotification: latestVersion }, { persistence: 'permanent' });
+    },
+  });
+ */
 
 function showDocsRoot() {
   const docsRoot = document.getElementById('docs-root')
@@ -40,28 +55,23 @@ type WithThemedPreviewProps = {
 export const WithFacelift = ({ children }: WithThemedPreviewProps) => {
   hideDocsRoot()
 
-  const addonState = useFaceliftSettings()
   const [showChildren, setShowChildren] = useState(false)
+  const addonState = useFaceliftState()
 
   useEffect(() => {
-    if (addonState) {
+    if (addonState.initialized) {
       setShowChildren(true)
       setTimeout(showDocsRoot, 0)
     }
   }, [addonState])
 
-  let Facelifted = (
-    <>
-      <PreviewStyles addonState={addonState} />
-      {showChildren && children}
-    </>
-  )
+  let Facelifted = <>{showChildren && children}</>
 
   if (addonState) {
     const {
       themeKey,
-      provider: stateProvider,
-      providerTheme: stateProviderTheme,
+      providerKey: stateProvider,
+      providerThemeKey: stateProviderTheme,
       themeVariant = 'light',
       themes,
       parameters,
@@ -69,43 +79,71 @@ export const WithFacelift = ({ children }: WithThemedPreviewProps) => {
     const {
       addProvider: _addProvider,
       autoThemeStory,
+      enhanceUi,
       provider: paramProvider,
       providerTheme: paramProviderTheme,
     } = parameters
+    Facelifted = (
+      <>
+        {enhanceUi && <PreviewStyles addonState={addonState} />}
+        {showChildren && children}
+      </>
+    )
+
     let addProvider = _addProvider
 
+    // Fallback for deprecated autoThemeStory
     if (!_addProvider && autoThemeStory) {
       addProvider = autoThemeStory
     }
 
+    if (!addProvider) {
+      return Facelifted
+    }
+
+    // Prepare provider values
     let providerKey = stateProvider || paramProvider
     let providerThemeKey = stateProviderTheme || paramProviderTheme
 
+    // If no provider is available - try the current themes provider if any
     if (!providerKey && themeKey) {
       const currentTheme = themes[themeKey]
       providerKey = currentTheme.provider
     }
 
+    // If no provider theme is available - try current themes provider theme if any
     if (!providerThemeKey && themeKey) {
       const currentTheme = themes[themeKey]
       providerThemeKey = currentTheme.providerTheme
     }
 
+    // If no providerTheme or providerThemeKey - no reason to continue
     if (!providerKey || !providerThemeKey) {
       return Facelifted
     }
 
-    const providerTheme = themes[providerThemeKey]
+    // Now check if provider theme is available
+    const providerTheme = themes[providerThemeKey] || {}
     const { instanciated, options } = providerTheme
+
+    // If no providerThemes intanciated or options are found - then no reason to continue
+    if (!instanciated || !options) {
+      return Facelifted
+    }
 
     const providerThemeOptions = options[themeVariant]
     const providerThemeInstanciated = instanciated[themeVariant]
+
+    // If no instanciated and options found of the variants - then no reason to continue
+    if (!providerThemeOptions || !providerThemeInstanciated) {
+      return Facelifted
+    }
 
     switch (providerKey) {
       case 'mui':
         Facelifted = (
           <>
-            <PreviewStyles addonState={addonState} />
+            {enhanceUi && <PreviewStyles addonState={addonState} />}
             {showChildren && providerThemeOptions && addProvider ? (
               <MuiThemeProvider
                 theme={createMuiTheme(providerThemeOptions as MuiThemeOptions)}
@@ -146,7 +184,7 @@ export const WithFacelift = ({ children }: WithThemedPreviewProps) => {
       case 'emotion':
         Facelifted = (
           <>
-            <PreviewStyles addonState={addonState} />
+            {enhanceUi && <PreviewStyles addonState={addonState} />}
             {showChildren && providerThemeInstanciated && addProvider ? (
               <EmotionThemeProvider
                 theme={providerThemeInstanciated}
@@ -163,7 +201,7 @@ export const WithFacelift = ({ children }: WithThemedPreviewProps) => {
       default:
         Facelifted = (
           <>
-            <PreviewStyles addonState={addonState} />
+            {enhanceUi && <PreviewStyles addonState={addonState} />}
             {showChildren && children}
           </>
         )