+++ src/constants.ts
@@ -1,7 +1,4 @@
 export const ADDON_ID = 'storybook-facelift'
 export const ADDON_PARAM_KEY = 'facelift'
-export const ADDON_EVENT_THEME_CHANGE = `${ADDON_ID}/theme-change`
-export const ADDON_EVENT_DOCS_RENDER = `${ADDON_ID}/docs-render`
-export const ADDON_THEME_SELECTOR_ICON = `${ADDON_ID}-theme-selector-icon`
-export const ADDON_THEME_SELECTOR = `${ADDON_ID}-theme-selector-tooltip`
+export const ADDON_EVENT_STATE_CHANGE = `${ADDON_ID}/state-change`
 export const ADDON_VARIANT_SELECTOR = `${ADDON_ID}-variant-selector`