+++ src/index.tsx
@@ -1,34 +1,29 @@
 // eslint-disable-next-line no-use-before-define
 import React, { useEffect, useState } from 'react'
-import addons, { makeDecorator } from '@storybook/addons'
-import { ADDON_EVENT_THEME_CHANGE, ADDON_PARAM_KEY } from './constants'
+import addons from '@storybook/addons'
 import { WithFacelift } from './components/WithFacelift'
+import { ADDON_EVENT_STATE_CHANGE } from './constants'
+import { defaultAddonState } from './defaults'
 
 import type { AddonState } from './typings/internal/state'
+import type { DecoratorFn } from '@storybook/react'
 
-export const useFaceliftSettings = () => {
-  const [addonSettings, setAddonSettings] = useState<AddonState | null>(null)
+export const useFaceliftState = () => {
+  const [faceliftState, setFaceliftState] = useState<AddonState>(defaultAddonState)
 
   useEffect(() => {
     const chan = addons.getChannel()
-
-    chan.on(ADDON_EVENT_THEME_CHANGE, setAddonSettings)
-
+    chan.on(ADDON_EVENT_STATE_CHANGE, setFaceliftState)
     return () => {
-      chan.off(ADDON_EVENT_THEME_CHANGE, setAddonSettings)
+      chan.off(ADDON_EVENT_STATE_CHANGE, setFaceliftState)
     }
-  })
+  }, [])
 
-  return addonSettings
+  return faceliftState
 }
 
-export const withFacelift = makeDecorator({
-  name: 'withFacelift',
-  parameterName: ADDON_PARAM_KEY,
-  skipIfNoParametersOrOptions: false,
-  wrapper: (storyFn, context) => {
-    return <WithFacelift>{storyFn(context)}</WithFacelift>
-  },
-})
+export const withFacelift: DecoratorFn = (storyFn) => {
+  return <WithFacelift>{storyFn()}</WithFacelift>
+}
 
-export * from './typings/internal/exposed'
+export type { AddonParameters as FaceliftParameters } from './typings/internal/parameters'