+++ src/typings/internal/state.ts
@@ -1,3 +1,4 @@
+import type { ThemeVars } from '@storybook/theming'
 import type {
   ThemeInstanciated,
   ThemeOptions,
@@ -5,16 +6,13 @@ import type {
   ThemeProviderType,
   ThemeVariantType,
 } from './common'
-import type { StorybookThemeOptions } from './exposed'
 import type { AddonParameters, ParamTheme } from './parameters'
 
 export type AddonStateOptionThemes = Partial<Record<ThemeVariantType, ThemeOptions>>
 
 export type AddonStateInstanciatedThemes = Partial<Record<ThemeVariantType, ThemeInstanciated>>
 
-export type AddonStateStorybookThemeOptions = Partial<
-  Record<ThemeVariantType, StorybookThemeOptions>
->
+export type AddonStateStorybookThemeOptions = Partial<Record<ThemeVariantType, ThemeVars>>
 
 export type AddonStateThemeTitles = Record<string, string>
 
@@ -26,7 +24,9 @@ export type AddonStateTheme = Omit<ParamTheme, 'variants'> & {
   storybook: AddonStateStorybookThemeOptions
 }
 
-export type AddonStateThemes = Record<string, AddonStateTheme>
+export type AddonStateThemes = {
+  [key: string]: AddonStateTheme
+}
 
 export type AddonState = {
   parameters: AddonParameters
@@ -34,8 +34,8 @@ export type AddonState = {
   themes: AddonStateThemes
   converters: AddonStateConverters
   themeTitles: AddonStateThemeTitles
-  providerTheme?: string
-  provider?: ThemeProviderType
+  providerThemeKey?: string
+  providerKey?: ThemeProviderType
   themeKey?: string
   themeVariant?: ThemeVariantType
 }