+++ CHANGELOG.md
@@ -1,2 +1,16 @@
-#### 6.8.6 (2021-05-26)
+### 6.9.0 (2021-05-29)
+
+##### Bug Fixes
+
+*  corrected npm version ([f315f22b](https://github.com/IgorSzyporyn/storybook-facelift/commit/f315f22b110dee36d910ef3ef7a6ea76bf968171))
+*  tweaked badgerui theme borders ([2772fa4b](https://github.com/IgorSzyporyn/storybook-facelift/commit/2772fa4bf377144a9d685ef71d4d253ab972775d))
+*  wrong tryped as proxied ThemeVars type was removed ([86c97ca8](https://github.com/IgorSzyporyn/storybook-facelift/commit/86c97ca82c4ec8aad416113d8617408973107bb4))
+*  exporting own parameters with own name ([a8f2ca80](https://github.com/IgorSzyporyn/storybook-facelift/commit/a8f2ca80c6d093f8c334a79283b6dbf15bb2a69b))
+*  cleaning up exposed types ([514d6cb9](https://github.com/IgorSzyporyn/storybook-facelift/commit/514d6cb9cc5904ad887fa4e5623e7771e308629c))
+*  all parameters now truly optional and zero-config with fallback to native theming in case of any problems ([b94ddd4c](https://github.com/IgorSzyporyn/storybook-facelift/commit/b94ddd4c6a80604d91436954c85d5636f759dfe5))
+*  lot of minor fixes ([fda44562](https://github.com/IgorSzyporyn/storybook-facelift/commit/fda44562fbddf55f6ebcc97ddbba511f9c4d17e7))
+
+##### Other Changes
+
+*  augmenting storybook parameters ([1fab91bb](https://github.com/IgorSzyporyn/storybook-facelift/commit/1fab91bb3e6d45cc0fe8207fa01a59fefcbded4c))